Add functional control structures
Previously, the code was written as friendly, procedural JS. This
is fine and totally cool, but I was pining for monads. This change
adds in the basic framework required to write this project in a
more functional way. I don't know whether it'll end up being used,
but it's a nice experiment.